% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_learner_factory.R
\name{create_learner}
\alias{create_learner}
\title{Creates an mlr3 learner}
\usage{
create_learner(config)
}
\arguments{
\item{config}{A list containing model configuration from \code{get_model_config()}.}
}
\value{
An mlr3 \code{Learner} object.
}
\description{
Dynamically creates an mlr3 learner object based on a model configuration.
The predict type is automatically set to "prob" to enable AUC calculation.
}
