% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_learner_factory.R
\name{create_multiple_learners}
\alias{create_multiple_learners}
\title{Creates a list of mlr3 learners}
\usage{
create_multiple_learners(configs)
}
\arguments{
\item{configs}{A list of configuration lists from \code{get_model_config()}.}
}
\value{
A list of mlr3 \code{Learner} objects.
}
\description{
Takes a list of model configurations and returns a list of corresponding
mlr3 learner objects.
}
