% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_learner_factory.R
\name{define_search_space}
\alias{define_search_space}
\title{Defines a hyperparameter search space}
\usage{
define_search_space(config)
}
\arguments{
\item{config}{A list containing model configuration.}
}
\value{
A \code{paradox::ParamSet} object representing the search space.
}
\description{
Returns a pre-defined hyperparameter search space for a given model configuration.
}
