% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{normalizePre}
\alias{normalizePre}
\title{Create the package default Questionnaire.}
\usage{
normalizePre(data, perform = T)
}
\value{
BdQuestionContainer object with default Questions
}
\description{
Create the package default Questionnaire.
}
\examples{

customQuestionnaire <- create_default_questionnaire()

}
