% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_benchmarking_engine.R
\name{plot_benchmark_summary}
\alias{plot_benchmark_summary}
\title{Visualize benchmark results}
\usage{
plot_benchmark_summary(bmr, measure = NULL)
}
\arguments{
\item{bmr}{The \code{BenchmarkResult} object from \code{run_model_benchmark}.}

\item{measure}{An optional \code{mlr3::Measure} to plot. If NULL, plots the primary
measure from the benchmark result.}
}
\value{
A \code{ggplot2} object.
}
\description{
Creates a boxplot comparing the performance of different learners from a
benchmark result.
}
