% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_results_analysis.R
\name{plot_tuning_summary}
\alias{plot_tuning_summary}
\title{Plot tuning summary}
\usage{
plot_tuning_summary(tuning_instance, type = "parameter")
}
\arguments{
\item{tuning_instance}{The finished tuning instance.}

\item{type}{The type of plot. Can be "parameter", "performance", or "history".}
}
\value{
A ggplot object.
}
\description{
Generates one of several standard diagnostic plots for a tuning instance.
}
