% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_tuning_engine.R
\name{run_hyperparameter_tuning}
\alias{run_hyperparameter_tuning}
\title{Run automated hyperparameter tuning}
\usage{
run_hyperparameter_tuning(
  task,
  learner,
  search_space,
  n_evals = 20,
  cv_folds = 5
)
}
\arguments{
\item{task}{The \code{mlr3} classification task.}

\item{learner}{The \code{mlr3} learner object.}

\item{search_space}{The \code{paradox::ParamSet} object.}

\item{n_evals}{The number of iterations for random search.}

\item{cv_folds}{The number of folds for cross-validation.}
}
\value{
A finished \code{mlr3tuning::TuningInstanceBatchSingleCrit} object containing all results.
}
\description{
The core tuning engine. It bundles a task, learner, and search space to find
the optimal hyperparameters using random search and cross-validation.
}
