% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{split_data}
\alias{split_data}
\title{Create the package default Questionnaire.}
\usage{
split_data(data)
}
\value{
BdQuestionContainer object with default Questions
}
\description{
Create the package default Questionnaire.
}
\examples{

customQuestionnaire <- create_default_questionnaire()

}
