% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_recommender.R
\name{suggest_learner}
\alias{suggest_learner}
\title{Suggest algorithms based on hybrid approach.}
\usage{
suggest_learner(dataset, type = "regression", predictor)
}
\value{
expected performance score for the dataset and algorithms
}
\description{
Suggest algorithms based on hybrid approach.
}
\examples{

val <- suggest_learner(data, "classsification", "Species")

}
