% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_recommender-manual.R
\name{suggest_learner_manual}
\alias{suggest_learner_manual}
\title{Suggest algorithms based on Decision tree method.}
\usage{
suggest_learner_manual(dataset, type = "regression", target)
}
\value{
expected performance score for the dataset and algorithm
}
\description{
Suggest algorithms based on Decision tree method.
}
\examples{

val <- suggest_learner_manual(data, "classsification", "Species")

}
