% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_recommender-learner.R
\name{suggest_learner_meta}
\alias{suggest_learner_meta}
\title{Suggest algorithms based on meta learninng technique.}
\usage{
suggest_learner_meta(dataset, type = "classification", predictor,
  algorithms)
}
\value{
expected performance score for the dataset and algorithm
}
\description{
Suggest algorithms based on meta learninng technique.
}
\examples{

val <- suggest_learner_meta(data, "classsification", "Species", "nnet")

}
