% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_final_training.R
\name{train_final_model}
\alias{train_final_model}
\title{Train the final model}
\usage{
train_final_model(learner, task, best_params)
}
\arguments{
\item{learner}{The \code{mlr3} learner object.}

\item{task}{The \code{mlr3} task.}

\item{best_params}{A list of the best hyperparameters from \code{summarize_tuning_results()}.}
}
\value{
The trained final learner object.
}
\description{
Applies the best hyperparameters to a learner and trains it on the full dataset.
}
