% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_recommender.R
\name{suggest_learner}
\alias{suggest_learner}
\title{Recommending learners based on dataset.}
\usage{
suggest_learner(dataset, type = "regression", predictor)
}
\arguments{
\item{dataset}{The dataset to train as a data frame}

\item{type}{Type of machine learning, either regression or classsification}

\item{predictor}{The target feature as character}
}
\value{
data frame of algorithms and thier ranks
}
\description{
Recommends machine learning algorithm best fitting for the given
dataset and learning type.
}
\examples{

ranking <- suggest_learner(iris, "classification", "Species")
ranking <- suggest_learner(cars, "regression", "mpg")

}
